compiler: move GC passes to the transform package
