+++ compiler/optimizer.go
@@ -137,8 +137,8 @@ func (c *Compiler) Optimize(optLevel, sizeLevel int, inlinerThreshold uint) erro
 	builder.Populate(modPasses)
 	modPasses.Run(c.mod)
 
-	hasGCPass := c.addGlobalsBitmap()
-	hasGCPass = c.makeGCStackSlots() || hasGCPass
+	hasGCPass := transform.AddGlobalsBitmap(c.mod)
+	hasGCPass = transform.MakeGCStackSlots(c.mod) || hasGCPass
 	if hasGCPass {
 		if err := c.Verify(); err != nil {
 			return errors.New("GC pass caused a verification failure")
